requirements.txt
Flask==3.0.0
gunicorn==21.2.0
psycopg[binary]==3.2.4
google-cloud-secret-manager==2.18.1
